backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Celery
redis==4.6.0
celery[redis]==5.3.4
flower==2.0.1

# S3/MinIO Storage
boto3==1.34.0
botocore==1.34.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
camelot-py[cv]==0.10.1
tabula-py==2.8.2

# OCR (optional)
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
Pillow==10.1.0

# AI/LLM
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Validation and Serialization
jsonschema==4.20.0
marshmallow==3.20.1

# Logging and Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0 